2019/15: oxygen system
Thought I'd go clever on this one and disassemble the intcode to find the answer easily.

I did extract the map and target position from the code, but in the end I still made a "proper" solution by running the simulation.

At least the dissassembly helped by showing that the map is a maze with no loops.

I figure the disassembler might come in handy later.