2023/20: pulse propagation
Hardest this year by far. I kind of had a hunch that I would have to find cycles and align them, but I got something wrong on the first try and then tangled up in some wild goose chase.

Finally had to resort to reddit for some hints which led to a working solution.