2019/21: Springdroid Adventure
Whoa, this was educational. I didn't even know where to begin at first, but bit by bit I realized that the solution is to try to jump as early as possible, i.e. if C is a hole, and none of the conditions that make that jump unsuccessful hold, then jump. Otherwise, if B is a hole, check the conditions that would prevent a jump, and finally if A is a hole, jump.

After sorting all this out, use boolean logic equivalences to simplify the expression to the point that it can be expressed with only two writable registers, and within the max number of instructions.